Draft for return size of image

+++ pkg/models/image.go
@@ -5,6 +5,6 @@ type Image struct {
 	ID      int64 `json:"id"`
 	Name    string `json:"name"`
 	Version string `json:"version"`
-	Size    uint64 `json:"size"`
+	Size    float64 `json:"size"`
 	Path    string `json:"path"`
 }
